requirements.txt
PyQt5==5.15.10
pyinstaller==6.3.0
requests>=2.31.0
python-dotenv>=1.0.0
markdown>=3.4.0

